Fix project cards image height
